# backend/docs/swagger.yaml
definitions:
...
  models.Thread:
    properties:
      body:
        type: string
      comments:
        type: integer
      createdAt:
        type: string
      deleted:
        type: boolean
      description:
        type: string
      id:
        type: integer
      image:
        type: string
      likes:
        type: integer
      moderationFlag:
        type: string
      shares:
        type: integer
      threadCorrections:
        items:
          $ref: '#/definitions/models.ThreadCorrection'
        type: array
      threadTags:
        items:
          $ref: '#/definitions/models.ThreadTag'
        type: array
      title:
        type: string
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/models.User'
      userID:
        type: integer
    type: object
  models.ThreadCorrection:
    properties:
      correction:
        type: string
      thread:
        $ref: '#/definitions/models.Thread'
      threadID:
        type: integer
    type: object
  models.ThreadTag:
    properties:
      tag:
        type: string
      thread:
        $ref: '#/definitions/models.Thread'
      threadID:
        type: integer
    type: object
  models.User:
    properties:
      aura:
        type: integer
      comments:
        type: integer
      createdAt:
        type: string
      id:
        type: integer
      passwordHash:
        type: string
      posts:
        type: integer
      profileImage:
        type: string
      threads:
        items:
          $ref: '#/definitions/models.Thread'
        type: array
      updatedAt:
        type: string
      username:
        type: string
    type: object
  thread_types.ThreadCreationForm:
    properties:
      body:
        type: string
      image:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
    type: object
  thread_types.ThreadResponse:
    properties:
      body:
        type: string
      comments:
        type: integer
      createdAt:
        type: string
      deleted:
        type: boolean
      description:
        type: string
      id:
        type: integer
      image:
        type: string
      liked:
        type: boolean
      likes:
        type: integer
      moderationFlag:
        type: string
      profileImage:
        type: string
      shares:
        type: integer
      threadCorrections:
        items:
          $ref: '#/definitions/models.ThreadCorrection'
        type: array
      threadTags:
        items:
          $ref: '#/definitions/models.ThreadTag'
        type: array
      title:
        type: string
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/models.User'
      userID:
        type: integer
      username:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /comment/:id/comments:
    get:
      consumes:
      - application/json
      description: Lists nested comment
      parameters:
      - description: commentID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Comments succesfully retrieved
          schema:
            additionalProperties:
              type: boolean
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Lists nested comment
      tags:
      - comments
  /comment/:id/like:
    delete:
      consumes:
      - application/json
      description: Unlikes a  comment
      parameters:
      - description: commentID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Comment successfully liked
          schema:
            additionalProperties:
              type: boolean
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Unlikes a comment
      tags:
      - comments
    post:
      consumes:
      - application/json
      description: Likes a  comment
      parameters:
      - description: commentID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Comment successfully liked
          schema:
            additionalProperties:
              type: boolean
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Likes a comment
      tags:
      - comments
...
  /thread/:id:
    delete:
      consumes:
      - application/json
      description: Deletes a thread
      parameters:
      - description: threadID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Thread successfully deleted
          schema:
            additionalProperties:
              type: boolean
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Deletes a thread
      tags:
      - threads
    get:
      consumes:
      - application/json
      description: Get a thread
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Thread successfully retrieved
          schema:
            $ref: '#/definitions/thread_types.ThreadResponse'
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a thread
      tags:
      - threads
    put:
      consumes:
      - application/json
      description: Edits a thread
      parameters:
      - description: Thread payload
        in: body
        name: thread
        required: true
        schema:
          $ref: '#/definitions/thread_types.ThreadCreationForm'
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Thread successfully edited
          schema:
            additionalProperties:
              type: boolean
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Edits a thread
      tags:
      - threads
...
  /thread/:id/comments:
    get:
      consumes:
      - application/json
      description: Lists direct comments
      parameters:
      - description: threadID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Comments succesfully retrieved
          schema:
            additionalProperties:
              type: boolean
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Lists direct comments
      tags:
      - comments
...
swagger: "2.0"